# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Service",
    "description": "FastAPI Service with Granian, uvloop, and SQLModel",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/payments/": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create payment invoice",
        "description": "Reserves a free Telegram account for 30m and executes the requested bot scenario to generate an invoice link. Cancels previous pending payment from the same user and reuses their account.",
        "operationId": "create_payment_api_v1_payments__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "List payments",
        "description": "Retrieve paginated list of payment transactions.",
        "operationId": "list_payments_api_v1_payments__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_PaymentRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/race": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Race all scenarios (SSE)",
        "description": "Runs all registered bot scenarios concurrently against all available accounts. Streams each successful payment link as a Server-Sent Event (SSE) as it is generated.\n\nSSE event types:\n- started: accounts locked\n- payment: link generated\n- error: scenario failed\n- done: race completed",
        "operationId": "create_payment_race_api_v1_payments_race_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/{payment_id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get payment details",
        "description": "Retrieve full details and status for a specific payment.",
        "operationId": "get_payment_api_v1_payments__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{payment_id}/paid": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Mark payment as paid",
        "description": "Reactive payment confirmation callback. Immediately frees the reserved Telegram account back to the pool.",
        "operationId": "mark_payment_paid_api_v1_payments__payment_id__paid_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PaymentCallback"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Callback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{payment_id}/cancel": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Cancel payment",
        "description": "Cancels a pending payment and immediately frees its Telegram account.",
        "operationId": "cancel_payment_api_v1_payments__payment_id__cancel_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{payment_id}/callback": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Generic payment status callback",
        "description": "Updates payment status via webhook and reactively handles account locking.",
        "operationId": "handle_payment_callback_api_v1_payments__payment_id__callback_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCallback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_PaymentRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaymentRead"
            },
            "type": "array",
            "title": "Items",
            "description": "List of records for the current page"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count of records matching criteria"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "Requested page size limit"
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Total number of available pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "PaginatedResponse[PaymentRead]"
      },
...
      "PaymentCallback": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PaymentStatus",
            "description": "Target status (e.g. paid, cancelled, failed)",
            "default": "paid"
          },
          "external_transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Transaction Id",
            "description": "External transaction reference ID"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "description": "Additional callback payload or provider response"
          }
        },
        "type": "object",
        "title": "PaymentCallback",
        "description": "Payload sent by webhooks or external providers to update payment status."
      },
      "PaymentCreate": {
        "properties": {
          "client_user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Client User Id",
            "description": "External client/user identifier requesting the payment",
            "examples": [
              "user_987654"
            ]
          },
          "scenario_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Scenario Id",
            "description": "Payment scenario identifier to execute",
            "examples": [
              "demo_bot"
            ]
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "description": "Monetary amount (must be positive)",
            "examples": [
              "500.00"
            ]
          },
          "currency": {
            "type": "string",
            "maxLength": 16,
            "title": "Currency",
            "description": "Currency code (e.g. RUB, USDT, USD, XTR)",
            "default": "RUB",
            "examples": [
              "RUB"
            ]
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "description": "Scenario-specific arguments (e.g. order_id, product_name)",
            "examples": [
              {
                "order_id": "ORD-12345"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "client_user_id",
          "scenario_id",
          "amount"
        ],
        "title": "PaymentCreate",
        "description": "Input payload for initiating a payment request."
      },
      "PaymentRaceCreate": {
        "properties": {
          "client_user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Client User Id",
            "description": "External client/user identifier requesting the payment",
            "examples": [
              "user_987654"
            ]
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "description": "Monetary amount (must be positive)",
            "examples": [
              "500.00"
            ]
          },
          "currency": {
            "type": "string",
            "maxLength": 16,
            "title": "Currency",
            "description": "Currency code (e.g. RUB, USDT, USD)",
            "default": "RUB",
            "examples": [
              "RUB"
            ]
          },
          "timeout_sec": {
            "type": "number",
            "maximum": 300.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout Sec",
            "description": "Maximum seconds to wait for all scenarios to complete",
            "default": 120.0,
            "examples": [
              120.0
            ]
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "description": "Additional scenario-specific arguments",
            "examples": [
              {}
            ]
          }
        },
        "type": "object",
        "required": [
          "client_user_id",
          "amount"
        ],
        "title": "PaymentRaceCreate",
        "description": "Input payload for initiating a concurrent multi-scenario payment race."
      },
      "PaymentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique payment ID (UUIDv7)"
          },
          "client_user_id": {
            "type": "string",
            "title": "Client User Id",
            "description": "Client/user identifier"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id",
            "description": "Assigned Telegram account ID"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id",
            "description": "Race batch ID (groups multiple scenario results)"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id",
            "description": "Executed payment scenario ID"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "description": "Payment amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Payment currency"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus",
            "description": "Current payment status"
          },
          "payment_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Link",
            "description": "Generated payment/invoice URL"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At",
            "description": "UTC expiration deadline"
          },
          "paid_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At",
            "description": "UTC payment completion timestamp"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At",
            "description": "UTC cancellation timestamp"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "description": "Scenario payload and responses"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_user_id",
          "scenario_id",
          "amount",
          "currency",
          "status",
          "payment_link",
          "expires_at",
          "paid_at",
          "cancelled_at",
          "meta",
          "created_at",
          "updated_at"
        ],
        "title": "PaymentRead",
        "description": "Response representation of a Payment entity."
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "cancelled",
          "expired",
          "failed"
        ],
        "title": "PaymentStatus",
        "description": "Lifecycle statuses for a payment transaction."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}